* Initial port of the debugger to GNU/Linux intel 64 bits - Thanks Gerardo!   - More work will follow :) * Add function frame definitions for x86 (ms, gcc), arm, ppc * Add more opcode descriptions in rsc adict:   - some arm floating point instructions   - basic powerpc instructions

+++ src/dbg/mem.h
@@ -16,7 +16,7 @@ typedef struct {
 
 	char *tag;
 	char *addr;
-	int size;
+	unsigned long size;
 
 	struct list_head list;
 
@@ -29,7 +29,7 @@ typedef struct {
 	unsigned long perms, perms_orig;
 	int flags;
 	char *bin;
-	int size;
+	unsigned long size;
 
 	struct list_head list;
